+++ packages/hooks/src/orderly/orderlyHooks.ts
@@ -39,6 +39,7 @@ export type {
 export { useChain } from "./useChain";
 export { useWithdraw } from "./useWithdraw";
 export { useDeposit } from "./useDeposit";
+export { useMaintenanceStatus } from "./useMaintenanceStatus";
 
 export { useHoldingStream } from "./useHoldingStream";
 export { useWalletSubscription } from "./useWalletSubscription";