+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @orderly/docs
 
+## 0.4.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+- Updated dependencies [e5a88d5f3]
+  - @orderly.network/react@1.5.10
+  - @orderly.network/hooks@1.5.10
+  - @orderly.network/types@1.5.10
+  - @orderly.network/web3-onboard@1.5.10
+
+## 0.4.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+- Updated dependencies
+  - @orderly.network/react@1.5.10-alpha.0
+  - @orderly.network/hooks@1.5.10-alpha.0
+  - @orderly.network/types@1.5.10-alpha.0
+  - @orderly.network/web3-onboard@1.5.10-alpha.0
+
 ## 0.4.9
 
 ### Patch Changes