+++ packages/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @orderly.network/core
 
+## 1.5.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+- Updated dependencies [e5a88d5f3]
+  - @orderly.network/types@1.5.10
+
+## 1.5.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+- Updated dependencies
+  - @orderly.network/types@1.5.10-alpha.0
+
 ## 1.5.9
 
 ### Patch Changes