+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/futures
 
+## 3.3.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+- Updated dependencies [e5a88d5f3]
+  - @orderly.network/types@1.5.10
+  - @orderly.network/utils@1.5.10
+
+## 3.3.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+- Updated dependencies
+  - @orderly.network/types@1.5.10-alpha.0
+  - @orderly.network/utils@1.5.10-alpha.0
+
 ## 3.3.9
 
 ### Patch Changes