+++ packages/walletconnect/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @orderly.network/web3-modal
 
+## 1.5.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+
+## 1.5.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+
 ## 1.5.9
 
 ### Patch Changes